Generate CRC-16 lookup table directly
Instead of including a static lookup table, we can just
generate it ourselves and save memory.  The lookup table only
holds 256 entires, so it's not CPU-intensive to compute.

We also get to remove 4k of code because we now just generate it
at runtime.  This also saves memory because previously the static
lookup table was being copied into our fast lookup table and never
used again.  Now we don't the one-time-use lookup table.

Also updated tests to compare against raw CRC16 computation, using
the lookup table, and using crc16speed.

+++ crcspeed.h
@@ -32,7 +32,7 @@
 #include <inttypes.h>
 
 typedef uint64_t (*crcfn64)(uint64_t, void *, uint64_t);
-typedef uint16_t (*crcfn16)(uint16_t, void *, uint16_t);
+typedef uint16_t (*crcfn16)(uint16_t, void *, uint64_t);
 
 /* CRC-64 */
 void crcspeed64little_init(crcfn64 fn, uint64_t table[8][256]);
